added in home page moved and fixed app.tsx
